% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_methods.R
\name{fit_cv}
\alias{fit_cv}
\alias{fit_naive}
\alias{fit_lasso}
\title{Fit methods}
\usage{
fit_cv(k_fold = 10)

fit_naive()

fit_lasso(k_fold = 10, alpha = 1, seed = 42)
}
\arguments{
\item{k_fold}{number of folds for crossvalidation.}

\item{alpha}{elasticnet mixing parameter passed to [glmnet::glmnet].
`alpha = 1` is the lasso penalty, and `alpha = 0` the ridge penalty.}

\item{seed}{seed for crossvalidation.}
}
\description{
Functions that do the actual fit.
}
