% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm2d.R
\name{lm2d}
\alias{lm2d}
\title{Regress a 2d field}
\usage{
lm2d(
  x,
  y,
  data = NULL,
  method = fit_cv(),
  max_eof = Inf,
  k_fold = 10,
  alpha = 1,
  seed = 42,
  verbose = FALSE
)
}
\arguments{
\item{x}{a formula to build the scalar field (see Details)}

\item{y}{vector}

\item{data}{data}

\item{method}{string with method}

\item{max_eof}{maximum number of principal components to use}

\item{k_fold}{number of folding for crossvalidation}

\item{alpha}{alpha number in lasso}

\item{seed}{seed used}

\item{verbose}{whether to talk a lot}
}
\value{
a list with elements:
\describe{
  \item{field}{the regression field}
  \item{fit}{the coefficients in EOF space}
  \item{summary}{a dataframe with r-squared, p.value and other aspects of the fit}
  \item{call}{the matched call}
  }
}
\description{
Regresses a 2D scalar field into one time series using the method descibed by
DelSole and Yang (2011).
}
\details{
The implementation from DelSole and Yang (2011) is `method = "cv"`. It uses
k-fold crossvalidation to select the first n principal components that
minimise the crossvalidation error when regressing `y` on `x`.

The `lasso` method, uses LASSO/ridge rergession regularization to
automatically filter the principal components. The `alpha` parameter controls
the relative weight of the LASSO and ridge penalities (`alpha = 1` is the
LASSO penalty and `alpha = 0` is the ridge penalty).

The `naive` method simply uses the first `max_eof` principal componets.
}
\references{
DelSole, T., & Yang, X. (2011). Field Significance of Regression Patterns. Journal of Climate, 24(19), 5094–5107. https://doi.org/10.1175/2011JCLI4105.1
}
